+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## [4.3.0] 16 / 10 / 2021
+## [4.3.0] 18 / 10 / 2021
 - Added most of textfield params to the phone input.
+- Added method to select the current national number from the controller
+- Changed how controllers worked under the hood
+- Fix an issue where a phone number could not start with its country code
+- uses phone_numbers_parser v4.1.0
+
 
 ## [4.2.0] 16 / 10 / 2021
 - [deprecated] PhoneValidator.invalid in favor of PhoneValidator.valid as the naming did not make sens and was backward.